3.2.0 (#835)
* CLI for whisperx and speechbrain transcription

* Optimizations for training acoustic models

* Switch to using miniforge for gha

* Fixing adaptation for older models

+++ montreal_forced_aligner/alignment/adapting.py
@@ -2,7 +2,6 @@
 from __future__ import annotations
 
 import logging
-import os
 import shutil
 import time
 from pathlib import Path
@@ -118,20 +117,20 @@ class AdaptingAligner(PretrainedAligner, AdapterMixin):
             IsmoothStatsAmDiagGmmFromModel(acoustic_model, self.mapping_tau, gmm_accs)
             objf_impr, count = transition_model.mle_update(transition_accs)
             logger.debug(
-                f"Transition model update: Overall {objf_impr/count} "
+                f"Transition model update: Overall {objf_impr / count} "
                 f"log-like improvement per frame over {count} frames."
             )
             objf_impr, count = acoustic_model.mle_update(
                 gmm_accs, update_flags_str="m", remove_low_count_gaussians=False
             )
             logger.debug(
-                f"GMM update: Overall {objf_impr/count} "
+                f"GMM update: Overall {objf_impr / count} "
                 f"objective function improvement per frame over {count} frames."
             )
             tot_like = gmm_accs.TotLogLike()
             tot_t = gmm_accs.TotCount()
             logger.debug(
-                f"Average Likelihood per frame = {tot_like/tot_t} " f"over {tot_t} frames."
+                f"Average Likelihood per frame = {tot_like / tot_t} " f"over {tot_t} frames."
             )
             write_gmm_model(str(final_mdl_path), transition_model, acoustic_model)
 
@@ -157,7 +156,7 @@ class AdaptingAligner(PretrainedAligner, AdapterMixin):
         """Current acoustic model path"""
         if self.current_workflow.workflow_type == WorkflowType.acoustic_model_adaptation:
             path = self.working_directory.joinpath("unadapted.alimdl")
-            if os.path.exists(path) and not getattr(self, "uses_speaker_adaptation", False):
+            if path.exists() and not getattr(self, "uses_speaker_adaptation", False):
                 return path
             return self.model_path
         return super().alignment_model_path
@@ -190,7 +189,7 @@ class AdaptingAligner(PretrainedAligner, AdapterMixin):
         """
         begin = time.time()
         log_directory = self.working_log_directory
-        os.makedirs(log_directory, exist_ok=True)
+        log_directory.mkdir(parents=True, exist_ok=True)
         self.acc_stats(alignment=False)
 
         if self.uses_speaker_adaptation:
@@ -204,19 +203,17 @@ class AdaptingAligner(PretrainedAligner, AdapterMixin):
         self.align()
         alignment_workflow = self.current_workflow
         self.create_new_current_workflow(WorkflowType.acoustic_model_adaptation)
-        for f in ["final.mdl", "final.alimdl"]:
+        for f in ["final.mdl", "final.alimdl", "tree", "lda.mat"]:
+            path = alignment_workflow.working_directory.joinpath(f)
+            new_path = self.working_directory.joinpath(f)
+            if f.startswith("final"):
+                new_path = new_path.with_stem("unadapted")
+            if not path.exists():
+                continue
             shutil.copyfile(
-                os.path.join(alignment_workflow.working_directory, f),
-                self.working_directory.joinpath(f).with_stem("unadapted"),
+                path,
+                new_path,
             )
-        shutil.copyfile(
-            os.path.join(alignment_workflow.working_directory, "tree"),
-            self.working_directory.joinpath("tree"),
-        )
-        shutil.copyfile(
-            os.path.join(alignment_workflow.working_directory, "lda.mat"),
-            self.working_directory.joinpath("lda.mat"),
-        )
         for j in self.jobs:
             old_paths = j.construct_path_dictionary(
                 alignment_workflow.working_directory, "ali", "ark"
@@ -224,28 +221,18 @@ class AdaptingAligner(PretrainedAligner, AdapterMixin):
             new_paths = j.construct_path_dictionary(self.working_directory, "ali", "ark")
             for k, v in old_paths.items():
                 shutil.copyfile(v, new_paths[k])
-        os.makedirs(self.align_directory, exist_ok=True)
+        self.align_directory.mkdir(parents=True, exist_ok=True)
         try:
             logger.info("Adapting pretrained model...")
             self.train_map()
             self.export_model(self.working_log_directory.joinpath("acoustic_model.zip"))
-            shutil.copyfile(
-                self.working_directory.joinpath("final.mdl"),
-                os.path.join(self.align_directory, "final.mdl"),
-            )
-            shutil.copyfile(
-                self.working_directory.joinpath("tree"),
-                os.path.join(self.align_directory, "tree"),
-            )
-            if os.path.exists(self.working_directory.joinpath("final.alimdl")):
+            for f in ["final.mdl", "final.alimdl", "tree", "lda.mat"]:
+                path = self.working_directory.joinpath(f)
+                if not path.exists():
+                    continue
                 shutil.copyfile(
-                    self.working_directory.joinpath("final.alimdl"),
-                    os.path.join(self.align_directory, "final.alimdl"),
-                )
-            if os.path.exists(self.working_directory.joinpath("lda.mat")):
-                shutil.copyfile(
-                    self.working_directory.joinpath("lda.mat"),
-                    os.path.join(self.align_directory, "lda.mat"),
+                    path,
+                    self.align_directory.joinpath(f),
                 )
             wf = self.current_workflow
             with self.session() as session:
@@ -317,6 +304,5 @@ class AdaptingAligner(PretrainedAligner, AdapterMixin):
         acoustic_model.add_model(self.working_directory)
         acoustic_model.add_model(self.phones_dir)
         if directory:
-            os.makedirs(directory, exist_ok=True)
-        basename, _ = os.path.splitext(output_model_path)
+            directory.mkdir(parents=True, exist_ok=True)
         acoustic_model.dump(output_model_path)
